pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aider-mcp"
version = "0.1.0"
authors = [
    {name = "Aider MCP Team", email = "info@example.com"},
]
description = "Model Context Protocol (MCP) server connecting Claude and other MCP clients to Aider"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "aider-chat>=0.26.0",
    "mcp>=0.1.0",
    "pydantic>=2.4.2",
    "typer>=0.9.0",
    "rich>=13.5.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "pre-commit>=3.3.3",
    "black>=23.7.0",
    "isort>=5.12.0",
    "mypy>=1.4.1",
]

[project.urls]
"Homepage" = "https://github.com/yourusername/aider-mcp"
"Bug Tracker" = "https://github.com/yourusername/aider-mcp/issues"

[project.scripts]
aider-mcp = "aider_mcp:main"

[tool.hatch.build.targets.wheel]
packages = ["src/aider_mcp"]

[tool.isort]
profile = "black"
line_length = 100

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false 